consumer/requirements.txt
kafka-python
psycopg2-binary
pytest